+++ HISTORY.md
@@ -1,5 +1,14 @@
 Version 1.0 (1.2021.09.14)
 ===================
-- Very basic set functions to run
-- Includes inverse of each function
-- includes all operations for f(x) and g(x)
+- Very basic set functions to run.
+- Includes inverse of each function.
+- includes all operations for f(x) and g(x).
+
+Version 1.0 (2.2021.09.15) 
+===================
+- Now focus is a website.
+- You can request how many functions you want.
+- There is a simple math bar.
+- For each function you add in, you can check the table of the values.
+- There is a line graph.
+- You can press "Draw all Values", it will draw all the functions on the graph.